Add files via upload
Excited to share my own Calculator as a milestone in my internship journey! 🚀
🌐 Project Title: [Calculator]
💼 Internship at: Codsoft
🔨 Tech Stack: HTML, CSS, Javascript

During this internship, I've had the opportunity to work on some amazing projects, and I'm proud to showcase the skills and knowledge I've gained. My blog is a collection of my best work, reflecting my growth and commitment to becoming a top-notch coder.